Move fixtures and fixturesDto to a fixtures folder
