+++ src/post/controller/__tests__/deletePost.test.ts
@@ -1,8 +1,8 @@
 import { NextFunction, Request, Response } from "express";
 import { Model } from "mongoose";
-import { luffyBentoPost } from "../../fixtures.js";
+import { luffyBentoPost } from "../../fixtures/fixtures.js";
 import { PostStructure } from "../../types.js";
-import { luffyBentoPostDto } from "../../dto/fixtures.js";
+import { luffyBentoPostDto } from "../../fixtures/fixturesDto.js";
 import PostController from "../PostController.js";
 import { PostRequest } from "../types.js";
 import ServerError from "../../../server/ServerError/ServerError.js";
