--- src/post/dto/fixtures.ts
@@ -1,15 +1,0 @@
-import { PostStructureDto } from "./types.js";
-
-export const luffyBentoPostDto: PostStructureDto = {
-  _id: "680d4b3e7a949d9fe3199b37",
-  author: "Sanji el Cocinero Galante",
-  content:
-    "Del East Blue a tu cocina: aprende a preparar el famoso bentō que Luffy ama. ¡Ideal para llevar al trabajo o picnic!",
-  imageAlt: "Bentō de One Piece",
-  imageUrl: "https://example.com/onepiece-bento.jpg",
-  publishDate: "2025-04-21",
-  tags: ["onepiece", "bento", "animefood"],
-  title: "Bentō pirata: la receta favorita de Luffy 🏴‍☠️",
-  smallImageUrl: "",
-  detailImageUrl: "",
-};
