+++ src/post/router/__tests__/addPostEndpoint.test.ts
@@ -3,7 +3,7 @@ import mongoose from "mongoose";
 import request from "supertest";
 import app from "../../../server/app.js";
 import connectToDatabase from "../../../database/connectToDatabase.js";
-import { tortillaBetanzosPostData } from "../../postDataFixtures.js";
+import { tortillaBetanzosPostData } from "../../fixtures/fixturesDto.js";
 import { responseBodyError, responseBodyPost } from "../../types.js";
 import Post from "../../model/Post.js";
